fix(ci): simplify npm publish workflow
Refactored GitHub Actions workflow for npm publishing:
- Removed redundant jobs and steps for platform-specific packages.
- Consolidated binary download and package creation into a single job.
- Added streamlined logic for handling binaries and publishing the main
package.
- Deleted outdated `install.js` script and replaced it with a new
version under `scripts/npm/install.js`.
- Introduced `package.json.template` for dynamic version substitution
during publishing.
