+++ scripts/npm/package.json.template
@@ -10,26 +10,19 @@
   },
   "files": [
     "install.js",
-    "bin/"
+    "bin/",
+    "binaries/"
   ],
   "keywords": ["openapi", "typescript", "validation", "io-ts", "zod", "schema", "dto"],
   "author": "Eliraz Kedmi",
   "license": "MIT",
   "repository": {
     "type": "git",
-    "url": "https://github.com/eliraz-refael/dtoForge.git"
+    "url": "git+https://github.com/eliraz-refael/dtoForge.git"
   },
   "homepage": "https://github.com/eliraz-refael/dtoForge#readme",
   "bugs": {
     "url": "https://github.com/eliraz-refael/dtoForge/issues"
-  },
-  "optionalDependencies": {
-    "dtoforge-darwin-x64": "${VERSION}",
-    "dtoforge-darwin-arm64": "${VERSION}",
-    "dtoforge-linux-x64": "${VERSION}",
-    "dtoforge-linux-arm64": "${VERSION}",
-    "dtoforge-win32-x64": "${VERSION}",
-    "dtoforge-win32-arm64": "${VERSION}"
   },
   "engines": {
     "node": ">=14"
